[REF] l10n_ve_pos_(mf_)self_order: split panel Kiosko base/MF
El panel de órdenes del Kiosko, el shell de debug y el loader
session_orders pasan a l10n_ve_pos_self_order: listar órdenes, recuperar
la factura de una orden pendiente y la persistencia de la cola no son
operaciones fiscales, y el ciclo de vida de la orden del Kiosko lo
gobierna ese módulo. Así la recuperación de factura desde el Kiosko
funciona aunque no esté instalado el módulo de máquina fiscal.

l10n_ve_pos_mf_self_order queda solo con los agregados fiscales (estado
pending_fiscal/complete, imprimir/reimprimir, pareo/estado de conexión),
enganchados por patch de componente + t-inherit sobre las anclas que
expone la plantilla de base. Un solo botón flotante de debug.

Traducciones:
- Se añade models/ir_http.py a l10n_ve_pos_mf_self_order
  (_get_translation_frontend_modules_name) para que su .po cargue en
  /website/translations (ruta pública del Kiosko). Sin esto, los strings
  fiscales salían en inglés aunque el .po estuviera completo.
- Se fusiona el es_VE.po de base (conservando las traducciones previas de
  la pantalla de identificación) y se completan las del panel/debug, con
  el nodo de texto del número fiscal reescrito para que sea traducible.
- Se corrige un string español hardcodeado sin _t() en printKioskFiscalInvoice.

Documentado en openspec (kiosk-orders-panel-to-base y
kiosk-fiscal-panel-extension).

References: Tarea 78767
