+++ l10n_ve_pos_mf_self_order/controllers/main.py
@@ -35,51 +35,3 @@ class L10nVePosMfSelfOrderController(PosSelfOrderController):
         return order.write_mf_invoice_data(
             mf_invoice_number, fiscal_machine, mf_reportz or False
         )
-
-    @http.route(
-        "/l10n_ve_pos_mf_self_order/kiosk/session_orders",
-        auth="public",
-        type="jsonrpc",
-        website=True,
-    )
-    def l10n_ve_kiosk_session_orders(self, access_token, limit=50):
-        """Órdenes de la sesión abierta de la caja, para el panel de órdenes
-        fiscales del Kiosko (persistencia real: no dependen de lo que haya en
-        memoria del cliente, que se pierde al iniciar una orden nueva o recargar).
-
-        Devuelve pos.order/line/payment/partner en el mismo formato que consume
-        ``connectNewData`` en el cliente, así el panel las lista y el builder
-        fiscal client-side las usa igual (líneas, impuestos, pago) para imprimir
-        o reimprimir la copia.
-        """
-        pos_config = self._verify_pos_config(access_token)
-        # Últimas órdenes de ESTA caja (cualquier sesión), no solo la sesión
-        # abierta: así el panel muestra también las de turnos anteriores para
-        # reimprimir copias o completar impresiones pendientes. Acotado por límite.
-        orders = (
-            pos_config.env["pos.order"]
-            .sudo()
-            .search(
-                [
-                    ("config_id", "=", pos_config.id),
-                    ("state", "in", ["paid", "done", "invoiced"]),
-                ],
-                order="id desc",
-                limit=int(limit) or 50,
-            )
-        )
-        if not orders:
-            return {}
-        env = pos_config.env
-        return {
-            "pos.order": env["pos.order"]._load_pos_self_data_read(orders, pos_config),
-            "pos.order.line": env["pos.order.line"]._load_pos_self_data_read(
-                orders.lines, pos_config
-            ),
-            "pos.payment": env["pos.payment"]._load_pos_self_data_read(
-                orders.payment_ids, pos_config
-            ),
-            "res.partner": env["res.partner"]._load_pos_self_data_read(
-                orders.partner_id, pos_config
-            ),
-        }
