+++ l10n_ve_pos_mf_self_order/models/__init__.py
@@ -1,3 +1,4 @@
+from . import ir_http
 from . import pos_config
 from . import pos_payment_method
 from . import account_tax
